Added Audio + Fix lives display
