+++ src/src_Phaser/GameConfig/PhaserConfig.jsx
@@ -7,8 +7,8 @@ const PhaserConfig = () => {
     useEffect(() => {
         const config = {
             type: Phaser.AUTO,
-            width: 1536,
-            height: 720,
+            width: window.innerWidth,
+            height: window.innerHeight,
             physics: {
                 default: 'arcade',
                 arcade: {
