+++ src/src_Phaser/Game_Objs/Platforms.jsx
@@ -5,6 +5,6 @@ export class Platforms extends Phaser.Physics.Arcade.StaticGroup {
         super(world, scene, children);
         this.scene = scene;
         this.number = number;
-        spriteArray.forEach(sprite => this.create(sprite.x, sprite.y, sprite.key));
+        spriteArray.forEach(sprite => this.create(sprite.x, sprite.y, sprite.key).setScale(1.5));
     }
 }
